requirements.txt
# Digital Sentience Server - Dependencies
# Optimizado para Render free tier (512MB RAM)

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Modelo LLM local (TinyLlama via llama.cpp)
# Nota: en Render se compila desde source, tarda ~5 min la primera vez
llama-cpp-python==0.2.77

# Base de datos vectorial (liviana, SQLite por debajo)
chromadb==0.5.3

# Embeddings semánticos (~90MB, rápido en CPU)
sentence-transformers==3.0.1

# Descargar modelos de HuggingFace
huggingface-hub==0.23.4

# Utilidades
pydantic==2.7.4
python-multipart==0.0.9   # para upload de archivos
psutil==6.0.0              # métricas de RAM/CPU

# Parsing de documentos
pypdf==4.2.0               # PDFs
python-docx==1.1.2         # .docx
